AMI resolution and overriding
If an AMI isn't specified then it will be resolved based
on region and instance type. It uses "resolvers" to do this.
Currebntly there are 2 resolvers:
- Default which uses the default EKS AMIs for a region
- GPU which for GPU instance types (p2 or p3) will use the
specific EKS GPU optimized image for a region.

This change also allows you to specify an AMI to use instead of
using resolution.

If using the GPU isntances then the NVIDIA device plugin will
still need to be installed manually:
https://github.com/NVIDIA/k8s-device-plugin
This will be an 'addon' in the future.

Issue #183